feat(ui): user role chip in profile dropdown, unified user menu across pages
- Add role chip (Viewer/Analyst/Admin) to the user dropdown in AppTopBar
- Extract dropdown into UserMenuButton component shared by AppTopBar and HubPage
- HubPage now has the full interactive dropdown instead of a static avatar + bare sign-out link
- Allow all roles to see their own recent activity on Hub (audit/recent now requires self:manage instead of audit:read; results are already scoped to caller.email)
- Fix AuditPage relative timestamps to use wall-clock time
