+++ backend/routes/audit.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Body, Depends, Header, HTTPException
+from fastapi import APIRouter, Body, Depends
 from pydantic import BaseModel
 from typing import List, Dict, Any, Optional
 from services.audit_service import (
@@ -8,24 +8,9 @@ from services.audit_service import (
 )
 from services.gemini_service import VisualizationConfig
 from services.rbac import require, Caller
-from services.azure_auth import extract_email_from_token
 
 router = APIRouter()
 
-
-def _require_caller_email(authorization: str) -> str:
-    """Resolve the caller's identity from the bearer token.
-
-    The token was already validated by Entra ID during the caller's OBO
-    exchange, so we only decode claims here. A missing/identity-less token is
-    rejected rather than waved through on the "Bearer " prefix alone.
-    """
-    if not authorization.startswith("Bearer "):
-        raise HTTPException(status_code=401, detail="Invalid token format")
-    email = extract_email_from_token(authorization[7:])
-    if not email:
-        raise HTTPException(status_code=401, detail="Caller identity missing")
-    return email
 
 
 class AuditQueryRequest(BaseModel):
@@ -61,12 +46,11 @@ class AuditEventItem(BaseModel):
 
 @router.get("/events", response_model=List[AuditEventItem])
 def audit_events(
-    authorization: str = Header(...),
     days: int = 90,
     limit: int = 1000,
     account: Optional[str] = None,
+    caller: Caller = Depends(require("audit:read")),
 ):
-    _require_caller_email(authorization)
     return get_audit_events(
         days=min(days, 365), limit=min(limit, 5000), account=account
     )
@@ -84,6 +68,6 @@ def query_audit_log(
 @router.get("/recent", response_model=List[RecentActivityItem])
 def recent_activity(
     limit: int = 10,
-    caller: Caller = Depends(require("audit:read")),
+    caller: Caller = Depends(require("self:manage")),
 ):
     return get_recent_activity(user_email=caller.email, limit=min(limit, 50))
